Part 2: Create different node sub types
If in the future it is likely that further node types may be added, then it would be beneficial to handle each of these node sub types separately so they contain only their own logic. Normally I would recommend using inheritance for this but in this javascript instance both the nodes simply require the same result method.

In a larger OOP example I would also recommend introducing interfaces for the Factory and Printer classes.